// DataDiscovery/ReportGenerationFront/src/app/core/models/LoginUser.ts
export interface LoginUser {
    createdOn: number;
    exp: number;
    expiriedOn: number;
    firstName: string;
    iat: number;
    lastName: string;
    roleName: string[];
    roleValue: number[];
    sub: string;
    userId: number;
    username: string;
    gender: number;
    designation: number;

}